Module Renaming
Enforced lowercase on all module names; removed import list in distil/active_learning_strategies/__init__.py
