+++ distil/active_learning_strategies/__init__.py
@@ -1,23 +1,4 @@
 # __init__.py
 # Author: Apurva Dani <apurvadani@gmail.com>
 
-from .fass import FASS
-from .entropy_sampling import EntropySampling
-from .entropy_sampling_dropout import EntropySamplingDropout
-from .random_sampling import RandomSampling
-from .least_confidence import LeastConfidence
-from .least_confidence_dropout import LeastConfidenceDropout
-from .margin_sampling import MarginSampling
-from .margin_sampling_dropout import MarginSamplingDropout
-from .core_set import CoreSet
-from .glister import GLISTER
-from .badge import BADGE
-from .gradmatch_active import GradMatchActive
-from .adversarial_bim import AdversarialBIM
-from .adversarial_deepfool import AdversarialDeepFool
-from .kmeans_sampling import KMeansSampling
-from .baseline_sampling import BaselineSampling
-from .bayesian_active_learning_disagreement_dropout import BALDDropout
-from .submod_sampling import SubmodSampling
-
 __version__ = '0.0.1'
