+++ distil/utils/submodular.py
@@ -5,7 +5,7 @@ import apricot
 
 from scipy.sparse import csr_matrix
 
-from .Similarity_mat import SimilarityComputation
+from .similarity_mat import SimilarityComputation
 
 
 class SubmodularFunction(SimilarityComputation):
